refector : string resource / ResourceProvider
In DetailActivity, i changed raw strings to string resource.
ResourceProvider is added to get string resource in ViewModel or somewhere. but this would be not a good way to use ResourceProvider in ViewModel because of the context.
but it would be also okay if we don't use ResourceProvider to get an app theme.